+++ health/service/src/test/scala/com/cerner/beadledom/health/resource/HealthResourceImplSpec.scala
@@ -1,79 +1,72 @@
 package com.cerner.beadledom.health.resource
 
-import com.cerner.beadledom.health.api.DependenciesResource
 import com.cerner.beadledom.health.dto.HealthDto
 import com.cerner.beadledom.health.internal.HealthChecker
-import com.cerner.beadledom.health.{TestModule1, TestModule2}
-import com.github.mustachejava.DefaultMustacheFactory
-import com.google.inject.{AbstractModule, Guice}
-import java.net.URI
-import javax.ws.rs.core.{UriBuilder, UriInfo}
-import org.mockito.ArgumentMatchers._
-import org.mockito.Mockito._
-import org.scalatest._
-import org.scalatest.mockito.MockitoSugar
+import com.cerner.beadledom.health.{HealthDependency, HealthServiceModule}
+import com.google.inject.multibindings.Multibinder
+import com.google.inject.{AbstractModule, Guice, Injector}
+import javax.ws.rs.core.UriInfo
 
 /**
- * Spec tests for HealthResourceImplSpec.
- *
- * @author Nimesh Subramanian
- */
-class HealthResourceImplSpec extends FunSpec with MustMatchers with MockitoSugar {
-  val mockUriInfo = mock[UriInfo]
-  val mockUriBuilder = mock[UriBuilder]
-  val url = URI.create("demoUrl")
+  * Spec tests for HealthResourceImplSpec.
+  *
+  * @author Nimesh Subramanian
+  */
 
-  when(mockUriInfo.getBaseUriBuilder()).thenReturn(mockUriBuilder)
-  when(mockUriBuilder.path(any[Class[DependenciesResource]])).thenReturn(mockUriBuilder)
-  when(mockUriBuilder.path(any[String])).thenReturn(mockUriBuilder)
-  when(mockUriBuilder.build()).thenReturn(url)
+class HealthResourceImplSpec(injector: Injector) extends BaseSpec {
 
   describe("HealthResourceImpl") {
     describe("#getPrimaryHealthCheck") {
       it("returns a dto with status 200") {
-        val mustacheFactory = new DefaultMustacheFactory("com/cerner/beadledom/health")
-        val healthModule = new AbstractModule {
+        val dependencyModule = new AbstractModule {
           override def configure(): Unit = {
-            install(new TestModule1)
+            val multibinder: Multibinder[HealthDependency] =
+              Multibinder.newSetBinder(binder(), classOf[HealthDependency])
+            multibinder.addBinding().toInstance(primaryHealthyDependency)
+            multibinder.addBinding().toInstance(ancillaryHealthyDependency)
+            install(new HealthServiceModule)
             bind(classOf[UriInfo]).toInstance(mockUriInfo)
           }
         }
 
-        val injector = Guice.createInjector(healthModule)
+        val injector = Guice.createInjector(dependencyModule)
         val checker = injector.getInstance(classOf[HealthChecker])
         val healthResource = new HealthResourceImpl(checker, mustacheFactory)
         val healthCheck = healthResource.getPrimaryHealthCheck
         val healthDto = healthCheck.getEntity.asInstanceOf[HealthDto]
 
-        healthDto.getMessage.get() must be ("Lombre is available")
-        healthDto.getStatus must be (200)
+        healthDto.getMessage.get() must be("myService is available")
+        healthDto.getStatus must be(200)
 
         val healthDependencies = healthDto.getDependencies.get()
 
-        healthDependencies.size() must be (2)
+        healthDependencies.size() must be(1)
       }
 
       it("returns a dto with status 503") {
-        val mustacheFactory = new DefaultMustacheFactory("com/cerner/beadledom/health")
-        val healthModule = new AbstractModule {
+        val dependencyModule = new AbstractModule {
           override def configure(): Unit = {
-            install(new TestModule2)
+            val multibinder: Multibinder[HealthDependency] =
+              Multibinder.newSetBinder(binder(), classOf[HealthDependency])
+            multibinder.addBinding().toInstance(primaryUnhealthyDependency)
+            multibinder.addBinding().toInstance(ancillaryUnhealthyDependency)
+            install(new HealthServiceModule)
             bind(classOf[UriInfo]).toInstance(mockUriInfo)
           }
         }
 
-        val injector = Guice.createInjector(healthModule)
+        val injector = Guice.createInjector(dependencyModule)
         val checker = injector.getInstance(classOf[HealthChecker])
         val healthResource = new HealthResourceImpl(checker, mustacheFactory)
         val healthCheck = healthResource.getPrimaryHealthCheck
         val healthDto = healthCheck.getEntity.asInstanceOf[HealthDto]
 
-        healthDto.getMessage.get() must be ("Lombre is unavailable")
-        healthDto.getStatus must be (503)
+        healthDto.getMessage.get() must be("artifactId is unavailable")
+        healthDto.getStatus must be(503)
 
         val healthDependencies = healthDto.getDependencies.get()
 
-        healthDependencies.size() must be (3)
+        healthDependencies.size() must be(1)
       }
     }
   }